Add files via upload
Project created to put into practice the content learned in the Rafa OutSystems Academy course. This project allows the user to register properties, evaluate them, and view the information. The project was created in a service module that contains the database structure, data manipulation methods and business rules. And a web module that contains the project's frontend and manipulation of information displayed on the screen.